requirements.txt
clean-fid
ndjson
numpy
matplotlib
pillow
torch
torchvision
tqdm
dotmap
pytorch-lightning
scikit-learn